pyproject.toml
[project]
name = "data-pipeline-mother-duck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==2.10.0",
    "duckdb>=1.4.4",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "python-dotenv>=1.2.1",
]
